Change register page to multistep
